Refactor computeResults into SimulationDriver
